Add implement custom wait and ReadyError
Currently, we use WaitUntilInstance(Exists|Running) in order to wait for
instances to be considered "ready". The issue with these functions is
that they only report that something went wrong, instead of why it went
wrong.

This adds a custom wait function that will return details about the
instance if it fails to enter running state.