vsphere: Answer questions attached to VM when possible.
If there are questions pending on a VM, the destroy operation
may fail repeatedly. It should be possible to answer questions
with a preferred answer.

The given preferred answer can either be the index of the choice
(as a string) or the text of the choice (case-independent).

There are also default options attached to questions and it may
be reasonable to answer all questions with the default unless
overwridden with a specific answer. However, this first change
just tries to answer only the predefined ones.

Note: There are no IDs for question types, which is why you must
store answers to questions keyed off of regular expressions.

Addresses issue #68